여행기록 게시판에서 이미지 클리하면 상세페이지로 넘어가는 기능 구현 완료
여행기록 작성페이지에서 이미지 업로드 및 ajax 활용한 도시 선택 구현예정

+++ src/main/java/com/team/semiTravelRecommend/model/dto/record/FileDTO.java
@@ -6,7 +6,7 @@ import lombok.Data;
 public class FileDTO {
 
     private int fileNo;
-    private int fileSize;
+    private long fileSize;
     private String originName;
     private String changeName;
     private String imgPath;
